+++ loxls/lsp/lifecyclemessages.go
@@ -49,6 +49,17 @@ func (h *Handler) initialize(params *protocol.InitializeParams[*initializationOp
 	}
 
 	h.initialized = true
+	var signatureHelpProvider *protocol.SignatureHelpOptions
+	if h.capabilities.GetTextDocument().GetSignatureHelp().GetContextSupport() {
+		signatureHelpProvider = &protocol.SignatureHelpOptions{
+			TriggerCharacters:   []string{"(", ","},
+			RetriggerCharacters: []string{")"},
+		}
+	} else {
+		signatureHelpProvider = &protocol.SignatureHelpOptions{
+			TriggerCharacters: []string{"(", ",", ")"},
+		}
+	}
 	return &protocol.InitializeResult{
 		Capabilities: &protocol.ServerCapabilities{
 			PositionEncoding: protocol.PositionEncodingKindUTF16,
@@ -64,6 +75,7 @@ func (h *Handler) initialize(params *protocol.InitializeParams[*initializationOp
 			HoverProvider: &protocol.BooleanOrHoverOptions{
 				Value: protocol.Boolean(true),
 			},
+			SignatureHelpProvider: signatureHelpProvider,
 			DefinitionProvider: &protocol.BooleanOrDefinitionOptions{
 				Value: protocol.Boolean(true),
 			},